requirements.txt
openai
python-nmap
requests
rich
pyfiglet
